[feature]Incluindo Pedido parametro da função,com os parâmetros do usuário

+++ src/cor-da-chita-api/cor-da-chita-api/Models/OrderDto.cs
@@ -1,5 +1,6 @@
 ﻿using MongoDB.Bson.Serialization.Attributes;
 using MongoDB.Bson;
+using System;
 
 namespace cor_da_chita_api.Models
 {
@@ -21,12 +22,12 @@ namespace cor_da_chita_api.Models
 
         public string ProductId { get; set; }
         public string ProductName { get; set; }
-        public double ProductPrice { get; set; }
+        public decimal ProductPrice { get; set; }
     }
 
     public class Freight
     {
-        public double TotalWithFreight { get; set; }
-        public double FreightValue { get; set; }
+        public decimal TotalWithFreight { get; set; }
+        public decimal FreightValue { get; set; }
     }
 }
